% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_alpha}
\alias{check_alpha}
\alias{check_theta}
\title{Check validity of copula parameters}
\usage{
check_alpha(alpha)

check_theta(theta)
}
\arguments{
\item{alpha}{Values of alpha to check.}

\item{theta}{Values of theta to check.}
}
\value{
An error if any theta or alpha is negative;
an invisible value otherwise. \code{NA} values do not throw an error.
}
\description{
Ensures input values are non-negative.
}
