% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize.R
\name{formals_to}
\alias{formals_to}
\title{Send arguments to a function after vectorizing}
\usage{
formals_to(.fn)
}
\arguments{
\item{.fn}{The function you want to send the recycled arguments to.}
}
\value{
The function \code{.fn} evaluated with the arguments given in
the encapsulating function.
}
\description{
When used within a (encapsulating) function, \code{formals_to}
recycles the inputs of the encapsulating function so that
they are vectors of the same length, and then sends these
updated arguments to some specified function.
}
