% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{y_pgamma}
\alias{y_pgamma}
\alias{y_dgamma}
\alias{y_qgamma}
\title{Gamma Distribution Links to C}
\usage{
y_pgamma(x, shape, scale = 1)

y_dgamma(x, shape, scale = 1)

y_qgamma(p, shape, scale = 1)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{shape, scale}{Vector of shape and scale parameters.}

\item{p}{Vector of probability levels.}
}
\value{
Vector of the evaluated representation of the Gamma distribution.
}
\description{
\code{pgamma()}, \code{dgamma()}, and \code{qgamma()} are defined natively in igcop,
written in C. The R functions named with a \code{y_} prefix link to the ones
in C.
}
