% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_gen.R
\name{ig_gen}
\alias{ig_gen}
\alias{ig_D1gen}
\alias{ig_geninv}
\title{Generating Function for the IG Copula Family}
\usage{
ig_gen(t, eta, k)

ig_D1gen(t, eta, k)

ig_geninv(p, eta, k, mxiter = 40, eps = 1e-12, bd = 5)
}
\arguments{
\item{t}{Vector of values >=1 to evaluate the function at.}

\item{eta}{Value of second argument of H_{k}, >0. This is allowed
to be a vector, except in \code{ig_geninv()}.}

\item{k}{Single numeric >1 corresponding to the \code{k} parameter
of the function.}

\item{w}{Vector of values in \link{0,1} to evaluate the inverse function at.}
}
\description{
\code{ig_gen} is the function itself, and \code{ig_geninv} is
its inverse; \code{ig_D1gen} is the first-argument
derivative.
}
