% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igamma.R
\name{igamma}
\alias{igamma}
\title{(Upper) Incomplete Gamma Function}
\usage{
igamma(s, x)
}
\arguments{
\item{s}{Scale parameter of the incomplete gamma function.
Vectorized, > 0.}

\item{x}{Lower bound of the incomplete gamma function.
Vectorized, > 0.}
}
\description{
Calculates the (upper) incomplete gamma function,
based on the Gamma cdf \code{pgamma}.
}
\details{
The (Upper) Incomplete Gamma Function is defined as
\deqn{\Gamma(s, x) = \int_x^{\infty} t^{s - 1} e^{-t} dt}.
}
\examples{
igamma(1:5, 1:5)
igamma(0.4, 1:5)
igamma(1:5, 0.4)

# Should be identical:
igamma(1:5, 0)
gamma(1:5)
}
