% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igcop.R
\name{igcond}
\alias{igcond}
\alias{igcondinv}
\title{Functions for IG copula 2|1 computations}
\usage{
igcond(t, k, eta)

igcondinv(p, k, eta, mxiter = 40, eps = 1e-06, bd = 5)
}
\arguments{
\item{t}{Vector of values to evaluate the assistant function at, >=1.}

\item{k}{Single numeric, greater than 1.}

\item{eta}{eta parameter. Vectorized to match \code{t} and \code{p}.}

\item{p}{Vector of values to evaluate the inverse assistant function at,
in \link{0,1}.}

\item{mxiter}{Maximum number of iterations to run the Newton-Raphson
algorithm for.}

\item{eps}{The Newton-Raphson
algorithm is stopped if all step sizes are below this value.}

\item{bd}{Largest acceptable step size in the Newton-Raphson algorithm;
steps larger than this will be reduced.}
}
\description{
Functions for IG copula 2|1 computations
}
