% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igcop-package.R
\docType{package}
\name{igcop}
\alias{igcop}
\title{igcop: Computational Tools for the IG and IGL Copula Families}
\description{
Compute distributional quantities for an
Integrated Gamma (IG) or IG Limit (IGL) copula, such
as a cdf and density, along with conditional quantities
such as the cdf, quantiles, and densities. Generate
data from a copula.
}
\section{Usage}{

Access copula quantities by starting with the \code{p}, \code{d}, \code{q}, or \code{r}
prefixes, followed by the copula name -- either \code{ig} or \code{igl}, or
their conditional versions, \code{condig} or \code{condigl}.
}

