% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igl_gen_inv_algo.R
\name{igl_gen_inv_algo}
\alias{igl_gen_inv_algo}
\title{Algorithm for computing inverse of IGL generator}
\usage{
igl_gen_inv_algo(p, alpha, mxiter = 20, eps = 1e-12, bd = 5)
}
\arguments{
\item{p}{Vector of values to evaluate the inverse function at, between
0 and 1 (inclusive).}

\item{alpha}{Parameter of the function, alpha > 0. Vectorized.}

\item{mxiter}{Maximum number of iterations to run the Newton-Raphson
algorithm for computing the inverse. Positive integer, defaults to 20.}

\item{eps}{The Newton-Raphson algorithm for computing an inverse will
stop if the step size is less than this small number.}

\item{bd}{The largest acceptable step size in the Newton-Raphson
algorithm. Step size is reduced if it exceeds this bound.}
}
\description{
Algorithm for computing inverse of IGL generator
}
