% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igl_kappa.R
\name{igl_kappa}
\alias{igl_kappa}
\title{Kappa version of generating function}
\usage{
igl_kappa(t, k)
}
\arguments{
\item{t}{Numeric argument of the kappa function. Vectorized.}

\item{k}{Parameter of the IGL generating function, \code{igl_gen()}. Vectorized.}
}
\value{
This function is a nearly direct application of the
\code{stats::pgamma()} function.
}
\description{
Defined as \eqn{\kappa(t) = \psi(t) - t \psi'(t)}, where
\eqn{\psi} is the IGL generating function \code{igl_gen()}.
}
