% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{igl_kappa_vec}
\alias{igl_kappa_vec}
\title{Functions generating the IG and IGL copulas}
\usage{
igl_kappa_vec(x, alpha)
}
\arguments{
\item{x}{Function argument}

\item{p}{Inverse function argument}

\item{eta, alpha}{Function parameters}
}
\description{
Kappa function and its relatives have prefix \code{igl_kappa};
Psi function and its relatives have prefix \code{igl_gen};
Interpolating function H with either kappa or psi has
\code{igl} prefix replaced with \code{interp}. Relatives of these functions:
suffix \code{inv} indicates inverse; suffix \code{D} represents function
derivative, and \code{D1} derivative with respect to the first argument.
. Suffix \verb{_vec} indicates that the entries must be vectors of
the same length; \verb{_single} means entries must be
scalars.
}
