% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_kappa_inv_algo.R
\name{interp_kappa_inv_algo}
\alias{interp_kappa_inv_algo}
\title{Algorithm: inverse of interpolator of kappa function}
\usage{
interp_kappa_inv_algo(p, eta, alpha, mxiter = 80, eps = 1e-12, bd = 5)
}
\arguments{
\item{p}{Vector of values between 0 and 1 to evaluate the inverse
function at.}

\item{eta}{Vector of values >0 of second argument of the
interpolating function.}

\item{alpha}{Vector of values >0 corresponding to the \eqn{alpha} parameter
of the IGL generating function.}

\item{mxiter}{Maximum number of iterations to run the Newton-Raphson
algorithm when computing inverse. Positive integer, default 20.}

\item{eps}{The Newton-Raphson algorithm for computing an inverse will
stop if the step size is less than this small number.}

\item{bd}{The largest acceptable step size in the Newton-Raphson
algorithm. Step size is reduced if it reaches this large.}
}
\description{
Algorithm: inverse of interpolator of kappa function
}
