% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcondig12_algo.R
\name{qcondig12_algo}
\alias{qcondig12_algo}
\title{Algorithm: Inverse of the 1|2 IG distribution function}
\usage{
qcondig12_algo(p, v, theta, alpha, mxiter = 80, eps = 1e-12, bd = 5)
}
\arguments{
\item{p}{Vector of quantile levels between 0 and 1
to evaluate a quantile function at.}

\item{v}{Vectors of values between 0 and 1 representing values of the first
and second copula variables.}

\item{theta}{Parameter of the IG copula family. Vectorized; >0.}

\item{alpha}{Parameter of the IG copula family. Vectorized; >0.}

\item{mxiter}{Maximum number of iterations to run the Newton-Raphson
algorithm when computing inverse. Positive integer, default 20.}

\item{eps}{The Newton-Raphson algorithm for computing an inverse will
stop if the step size is less than this small number.}

\item{bd}{The largest acceptable step size in the Newton-Raphson
algorithm. Step size is reduced if it reaches this large.}
}
\description{
Algorithm: Inverse of the 1|2 IG distribution function
}
