% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{.u}
\alias{.u}
\alias{.v}
\alias{.cpar}
\alias{.theta}
\alias{.alpha}
\title{Test data}
\format{
Everything is a numeric vector, except \code{.cpar}, which is a list
of bivariate numeric vectors. \code{.u} and \code{.v} are of matching
length; the rest are not intended to have matching lengths.

An object of class \code{numeric} of length 17.

An object of class \code{list} of length 25.

An object of class \code{numeric} of length 14.

An object of class \code{numeric} of length 11.
}
\usage{
.u

.v

.cpar

.theta

.alpha
}
\description{
Internal data used in the test scripts. \code{.u} and \code{.v} are vectors of
matching length containing values between 0 and 1 and reasonably cover
the unit square. \code{.cpar} is a list of IG copula parameter pairs
\code{c(theta, alpha)}, and \code{.theta} and \code{.alpha} are the corresponding
(unique) individual values.
}
\keyword{datasets}
