% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unload.R
\name{.onUnload}
\alias{.onUnload}
\title{Clean up DLL}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{Argument}
}
\description{
As recommended in the "Compiled Code" chapter of the book
"R packages" (Version 2) by Hadley Wickham and Jenny Bryan.
}
