% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig.R
\name{pcondig21}
\alias{pcondig21}
\alias{qcondig21}
\alias{qcondig}
\alias{pcondig}
\alias{qcondig12}
\alias{pcondig12}
\alias{dig}
\alias{logdig}
\alias{pig}
\alias{rig}
\title{IG Copula Family Functions}
\usage{
pcondig21(v, u, cpar)

qcondig21(tau, u, cpar)

qcondig(tau, u, cpar)

pcondig(v, u, cpar)

qcondig12(tau, v, cpar, mxiter = 80, eps = 1e-12, bd = 5)

pcondig12(u, v, cpar)

dig(u, v, cpar)

logdig(u, v, cpar)

pig(u, v, cpar)

rig(n, cpar)
}
\arguments{
\item{u, v}{Vectors of values between 0 and 1 representing values of the first
and second copula variables.}

\item{cpar}{Vector of length 2 corresponding to the copula
parameters \code{theta>0} and \code{alpha>0}, respectively.}

\item{tau}{Vector of quantile levels between 0 and 1
to evaluate a quantile function at.}

\item{mxiter}{Maximum number of iterations to run the Newton-Raphson
algorithm when computing inverse. Positive integer, default 20.}

\item{eps}{The Newton-Raphson algorithm for computing an inverse will
stop if the step size is less than this small number.}

\item{bd}{The largest acceptable step size in the Newton-Raphson
algorithm. Step size is reduced if it reaches this large.}

\item{n}{Positive integer. Number of observations to randomly draw.}
}
\value{
Numeric vector of length equal to the length of the input vector(s).
}
\description{
Functions related to the IG copula family, denoted  by \code{'ig'}.
}
\note{
Inputting two vectors greater than length 1 is allowed, if they're
the same length.
Also, \code{qcondig21} and \code{pcondig21} are the same as
\code{qcondig} and \code{pcondig} -- they're the distributions of
variable 2 given 1.
}
\examples{
set.seed(1)
u <- runif(10)
v <- runif(10)
pig(u, v, cpar = c(5, 1))
dig(u, v, cpar = c(2, 2))
logdig(u, v, cpar = c(2, 2))
pcondig21(v, u, cpar = c(3, 6))
qcondig21(v, u, cpar = c(3, 6))
pcondig12(u, v, cpar = c(3, 6))
qcondig12(u, v, cpar = c(3, 6))
rig(10, cpar = c(3, 3))
}
