% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igcop.R
\name{pcondigcop21}
\alias{pcondigcop21}
\alias{qcondigcop21}
\alias{qcondigcop}
\alias{pcondigcop}
\alias{qcondigcop12}
\alias{pcondigcop12}
\alias{digcop}
\alias{logdigcop}
\alias{pigcop}
\alias{rigcop}
\title{IG Copula Family Functions}
\usage{
pcondigcop21(v, u, cpar)

qcondigcop21(tau, u, cpar)

qcondigcop(tau, u, cpar)

pcondigcop(v, u, cpar)

qcondigcop12(tau, v, cpar, mxiter = 80, eps = 1e-12)

pcondigcop12(u, v, cpar)

digcop(u, v, cpar)

logdigcop(u, v, cpar)

pigcop(u, v, cpar)

rigcop(n, cpar)
}
\arguments{
\item{u, v}{Vectors of values between 0 and 1 representing values of the first
and second copula variables.}

\item{cpar}{Vector of length 2 corresponding to the copula
parameters \code{theta>0} and \code{alpha>0}, respectively.}

\item{tau}{Vector of quantile levels between 0 and 1
to evaluate a quantile function at.}

\item{n}{Positive integer. Number of observations to randomly draw.}
}
\value{
Numeric vector of length equal to the length of the input vector(s).
}
\description{
Functions related to the IG copula family, denoted  by \code{'igcop'}.
}
\note{
Inputting two vectors greater than length 1 is allowed, if they're
the same length.
Also, \code{qcondigcop21} and \code{pcondigcop21} are the same as
\code{qcondigcop} and \code{pcondigcop} -- they're the distributions of
variable 2 given 1.
}
