% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igl_gen.R
\name{igl_gen}
\alias{igl_gen}
\alias{igl_gen_D}
\alias{igl_gen_DD}
\alias{igl_gen_inv_algo}
\alias{igl_gen_inv}
\title{Generating function for the IGL copula family}
\usage{
igl_gen(x, k)

igl_gen_D(x, k)

igl_gen_DD(x, k)

igl_gen_inv_algo(p, k, mxiter = 20, eps = 1e-12, bd = 5)

igl_gen_inv(p, k, mxiter = 20, eps = 1e-12, bd = 5)
}
\arguments{
\item{x}{Vector of values >=0 to evaluate the function at, .}

\item{k}{Parameter of the function, k > 1. Vectorized.}

\item{p}{Vector of values to evaluate the inverse function at, between
0 and 1 (inclusive).}
}
\description{
\code{igl_gen} is the function itself, and \code{igl_gen_inv} is
its inverse; \code{igl_gen_D} is the
derivative; and \code{igl_gen_DD} is the second derivative.
}
\details{
Function arguments and parameters are vectorized, except
for the algorithms (marked by \verb{_algo}).
}
\examples{
## Some examples of evaluating the functions.
arg <- c(0, 0.5, 3, Inf, NA)
#igl_gen(arg, k = 2)
#igl_gen_D(arg, k = 1.2)
#igl_gen_D(arg, k = 2)
#igl_gen_D(arg, k = 3)
#igl_gen_inv(c(0, 0.5, 1), k = 1.5)

## Visual
#foo <- function(u) igl_gen_inv(u, k = 1.5)
#curve(foo)
}
