% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igl_gen.R, R/y_igl_gen_inv.R
\name{igl_gen}
\alias{igl_gen}
\alias{igl_gen_D}
\alias{y_igl_gen_inv}
\title{Generating function for the IGL copula family}
\usage{
igl_gen(x, alpha)

igl_gen_D(x, alpha)

y_igl_gen_inv(p, alpha)
}
\arguments{
\item{x}{Vector of values >=0 to evaluate the function at.}

\item{alpha}{Parameter of the function, alpha > 0. Vectorized.}

\item{p}{Vector of values to evaluate the inverse function at, between
0 and 1 (inclusive).}
}
\description{
\code{igl_gen} is the function itself; \code{igl_gen_inv} is
its inverse; and \code{igl_gen_D} is the
derivative. Function arguments and parameters are vectorized.

\code{igl_gen} is the function itself; \code{igl_gen_inv} is
its inverse; and \code{igl_gen_D} is the
derivative. Function arguments and parameters are vectorized.
}
