% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolators.R
\name{interp_gen}
\alias{interp_gen}
\alias{interp_kappa}
\alias{interp_gen_D1}
\alias{interp_gen_inv}
\title{Interpolating Functions}
\usage{
interp_gen(t, eta, k)

interp_kappa(t, eta, k)

interp_gen_D1(t, eta, k)

interp_gen_inv(p, eta, k, mxiter = 40, eps = 1e-12, bd = 5)
}
\arguments{
\item{t}{Vector of values >=1 to evaluate the interpolating function at.}

\item{eta}{Value of second argument of the interpolating function, >0.
This is allowed to be a vector, except in the inverse functions.}

\item{k}{Single numeric >1 corresponding to the \eqn{k} parameter
of the IGL generating function.}

\item{p}{Vector of values in \link{0,1} to evaluate the inverse function at.}
}
\description{
These interpolating functions, denoted \eqn{H} in the vignette,
depend on a generating function (of a DJ copula).
\code{interp_gen()} uses the IGL generating function \eqn{\Psi_k};
\code{interp_kappa()} uses the "kappa version" of that same function.
}
\details{
Appending \verb{_inv} to the function name indicates inverse with
respect to the first argument. Appending \verb{_D1} indicates
derivative with respect to the first argument.
}
