% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_igl_kappa.R
\name{y_igl_kappa}
\alias{y_igl_kappa}
\alias{y_igl_kappa_D}
\alias{y_igl_kappa_inv}
\title{Kappa version of generating function}
\usage{
y_igl_kappa(x, alpha)

y_igl_kappa_D(x, alpha)

y_igl_kappa_inv(p, alpha)
}
\arguments{
\item{x}{Numeric argument of the kappa function. Vectorized.}

\item{alpha}{Parameter of the IGL generating function, \code{igl_gen()}, >0.
Vectorized.}

\item{p}{Numeric argument of the inverse function. Vectorized.
Between 0 and 1, inclusive.}
}
\description{
The kappa version of a generating function \eqn{\psi} is
defined as \eqn{\kappa(x) = \psi(x) + x \psi'(x)}.
\code{igl_kappa} takes \eqn{\psi} to be the IGL generating
function, \code{igl_gen()}.
}
\details{
\code{igl_kappa_D()} is the derivative; \code{igl_kappa_inv()} is the inverse.
}
