% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igl_kappa.R
\name{igl_kappa}
\alias{igl_kappa}
\alias{igl_kappa_D}
\alias{igl_kappa_inv}
\title{Kappa version of generating function}
\usage{
igl_kappa(x, alpha)

igl_kappa_D(x, alpha)

igl_kappa_inv(p, alpha)
}
\arguments{
\item{x}{Numeric argument of the kappa function. Vectorized.}

\item{alpha}{Parameter of the IGL generating function, \code{igl_gen()}, >0. Vectorized.}

\item{p}{Numeric argument of the inverse function. Vectorized. Between 0 and 1.}
}
\description{
The kappa version of a generating function \eqn{\psi} is
defined as \eqn{\kappa(x) = \psi(x) + x \psi'(x)}.
\code{igl_kappa} takes \eqn{\psi} to be the IGL generating
function, \code{igl_gen()}.
}
\details{
\code{igl_kappa_D()} is the derivative.
}
