% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_square.R
\name{trim_square}
\alias{trim_square}
\title{Trim values to (0, 1)}
\usage{
trim_square(u, eps = 1e-06, lower = TRUE, upper = TRUE)
}
\arguments{
\item{u}{Vector of values.}

\item{eps}{Distance from 0 and 1 for which to truncate.}

\item{lower, upper}{Logical; should the lower bound (0) be considered
in the truncation? Should the upper bound (1)?}
}
\value{
The vector \code{u} with its values possibly truncated.
}
\description{
Values between 0 and 1 that are to close to either 0 or 1
are replaced with epsilon or 1-epsilon, respectively.
}
\note{
This function is used to trim copula inputs, and
is intended to be a temporary fix until functions like
igcop's in-house \code{qgamma} and inverse functions have their
accuracies improved.
}
